requirements.txt
Flask
requests
python-dotenv
gunicorn
prometheus-client